Move the tests of auth
